requirements.txt
streamlit
openai
google-auth-oauthlib
google-auth
google-api-python-client